Abstract realm and emoji API details.
